Restructured nodejs modules
