+++ src/app/components/installations/CameraComponent.jsx
@@ -341,7 +341,7 @@ export default class CameraComponent extends React.Component {
       type: "Installation"
     }; // if we need to send parameters to the server request
     var ft = new FileTransfer();
-    ft.upload(fileURI, encodeURI("http://" + IP + "/upload"), win, fail, options);
+    ft.upload(fileURI, encodeURI("http://" + IP + "/common/upload"), win, fail, options);
   }
 
   desktopUploadClickHandler() {
@@ -374,7 +374,7 @@ export default class CameraComponent extends React.Component {
 
       }
     };
-    httpRequest.open('POST', "http://" + IP + '/upload', true);
+    httpRequest.open('POST', "http://" + IP + '/common/upload', true);
     httpRequest.send(fd);
 
     this.setState({
